add delete falsy leads in brevo one shot script

+++ back/src/domains/marketing/adapters/establishmentMarketingGateway/BrevoEstablishmentMarketingGateway.ts
@@ -105,9 +105,13 @@ export class BrevoEstablishmentMarketingGateway
   async delete(contactEmail: Email): Promise<void> {
     const existingContact = await this.getContactByEmail(contactEmail);
 
-    return !existingContact ||
-      (existingContact?.lists.length === 1 &&
-        existingContact?.lists.at(0) === this.#establishmentContactListId)
+    if (
+      !existingContact ||
+      !existingContact.lists.includes(this.#establishmentContactListId)
+    )
+      return;
+
+    return existingContact.lists.length === 1
       ? this.#deleteContactByEmail(contactEmail)
       : this.#createContact({
           attributes: {
